feat: lst.h - a header only for lst functions
