+++ inc/minishell.h
@@ -6,7 +6,7 @@
 /*   By: bmugnol- <bmugnol-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/12 19:40:30 by llopes-n          #+#    #+#             */
-/*   Updated: 2022/07/15 04:05:12 by bmugnol-         ###   ########.fr       */
+/*   Updated: 2022/07/15 04:44:11 by bmugnol-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,6 +19,7 @@
 # include <readline/history.h>
 # include "libft.h"
 # include <fcntl.h>
+# include "lst.h"
 
 typedef struct s_glo
 {
@@ -26,22 +27,13 @@ typedef struct s_glo
 	char	**args;
 }	t_glo;
 
-typedef struct s_node
-{
-	char			*name;
-	char			*value;
-	struct s_node	*next;
-}	t_node;
-
 typedef enum e_bool
 {
 	TRUE,
 	FALSE
 }	t_bool;
 
-t_node	*shlst_new(char *name, char *value);
-void	shlst_addback(t_node **lst, t_node *new);
-t_node	*set_node(void);
-t_node	*find_var(char *name, t_node *lst);
 char	*get_comman(char *comman, t_node *env_lst);
+t_node	*set_node(void);
+
 #endif
